move application pool into a separate file, save session/application status change in a map directly, do not use channel, it could block when channel is full
