+++ cmd/simulation/node/app/snode/node_actor.go
@@ -408,9 +408,8 @@ func (n *SimulationNodeActor) onSessionOpenCommand(msg *fornaxgrpc.SessionOpen) 
 			sess.ResourceVersion = fmt.Sprint(revision)
 			fpod.Sessions[sessId] = fsess
 			fsess.Session.Status.SessionStatus = fornaxv1.SessionStatusAvailable
-			fsess.Session.Status.AvailableTime = util.NewCurrentMetaTime()
-			fsess.Session.Status.AvailableTimeMicro = time.Now().UnixMicro()
 			n.notify(n.fornoxCoreRef, session.BuildFornaxcoreGrpcSessionState(revision, fsess))
+			klog.InfoS("Opened session", "session", msg.SessionIdentifier, "pod", msg.PodIdentifier, "node", n.node.V1Node.Name)
 		}()
 	}
 	return nil
@@ -433,8 +432,8 @@ func (n *SimulationNodeActor) onSessionCloseCommand(msg *fornaxgrpc.SessionClose
 				revision := n.incrementNodeRevision()
 				fsess.Session.ResourceVersion = fmt.Sprint(revision)
 				fsess.Session.Status.SessionStatus = fornaxv1.SessionStatusClosed
-				fsess.Session.Status.CloseTime = util.NewCurrentMetaTime()
 				n.notify(n.fornoxCoreRef, session.BuildFornaxcoreGrpcSessionState(revision, fsess))
+				klog.InfoS("Closed session", "session", msg.SessionIdentifier, "pod", msg.PodIdentifier, "node", n.node.V1Node.Name)
 			}()
 		} else {
 			return fmt.Errorf("Session: %s does not exist, can not terminate session", sessId)
