+++ pkg/fornaxcore/application/application_pod.go
@@ -64,159 +64,6 @@ func NewApplicationPod(podName string, state ApplicationPodState) *ApplicationPo
 	}
 }
 
-type ApplicationPodSummary struct {
-	totalCount    int32
-	pendingCount  int32
-	deletingCount int32
-	idleCount     int32
-	occupiedCount int32
-}
-
-func (pool *ApplicationPool) summaryPod(podManager ie.PodManagerInterface) ApplicationPodSummary {
-	pool.mu.RLock()
-	defer pool.mu.RUnlock()
-	summary := ApplicationPodSummary{}
-	summary.pendingCount = int32(len(pool.pods[PodStatePending]))
-	summary.deletingCount = int32(len(pool.pods[PodStateDeleting]))
-	summary.occupiedCount = int32(len(pool.pods[PodStateAllocated]))
-	summary.idleCount = int32(len(pool.pods[PodStateIdle]))
-	summary.totalCount = summary.pendingCount + summary.deletingCount + summary.idleCount + summary.occupiedCount
-
-	return summary
-}
-
-func (pool *ApplicationPool) getPodSessions(podName string) []*fornaxv1.ApplicationSession {
-	sessions := []*fornaxv1.ApplicationSession{}
-	pool.mu.RLock()
-	defer pool.mu.RUnlock()
-	pod := pool._getPodNoLock(podName)
-	if pod == nil {
-		return sessions
-	}
-	for k := range pod.sessions {
-		if sess := pool._getSessionNoLock(k); sess != nil {
-			sessions = append(sessions, sess)
-		}
-	}
-	return sessions
-}
-
-func (pool *ApplicationPool) getPod(podName string) *ApplicationPod {
-	pool.mu.RLock()
-	defer pool.mu.RUnlock()
-	return pool._getPodNoLock(podName)
-}
-
-func (pool *ApplicationPool) _getPodNoLock(podName string) *ApplicationPod {
-	for _, pods := range pool.pods {
-		if p, found := pods[podName]; found {
-			return p
-		}
-	}
-	return nil
-}
-
-func (pool *ApplicationPool) addOrUpdatePod(podName string, podState ApplicationPodState, sessionIds []string) *ApplicationPod {
-	pool.mu.Lock()
-	defer pool.mu.Unlock()
-	return pool._addOrUpdatePodNoLock(podName, podState, sessionIds)
-}
-
-func (pool *ApplicationPool) _addOrUpdatePodNoLock(podName string, podState ApplicationPodState, sessionIds []string) *ApplicationPod {
-	for _, pods := range pool.pods {
-		// found in a state map, update
-		if p, f := pods[podName]; f {
-			for _, v := range sessionIds {
-				p.sessions[v] = true
-			}
-			if p.state == podState {
-				return p
-			} else {
-				p.state = podState
-				pool.pods[podState][podName] = p
-				delete(pods, podName)
-				return p
-			}
-		}
-	}
-
-	// not found, add it
-	p := NewApplicationPod(podName, podState)
-	for _, v := range sessionIds {
-		p.sessions[v] = true
-	}
-	pool.pods[podState][podName] = p
-	return p
-}
-
-func (pool *ApplicationPool) deletePod(podName string) {
-	pool.mu.Lock()
-	defer pool.mu.Unlock()
-	for _, v := range pool.pods {
-		delete(v, podName)
-	}
-}
-
-func (pool *ApplicationPool) getSomeIdlePods(num int) []*ApplicationPod {
-	pool.mu.RLock()
-	defer pool.mu.RUnlock()
-	pods := []*ApplicationPod{}
-	for _, v := range pool.pods[PodStateIdle] {
-		if len(pods) == num {
-			break
-		}
-		pods = append(pods, v)
-	}
-	return pods
-}
-
-func (pool *ApplicationPool) activeApplicationPodLength() (occupiedPods, pendingPods, idlePods int) {
-	pool.mu.RLock()
-	defer pool.mu.RUnlock()
-	pendingPods = len(pool.pods[PodStatePending])
-	idlePods = len(pool.pods[PodStateIdle])
-	occupiedPods = len(pool.pods[PodStateAllocated])
-	return occupiedPods, pendingPods, idlePods
-}
-
-func (pool *ApplicationPool) podsOfState(state ApplicationPodState) []*ApplicationPod {
-	pool.mu.RLock()
-	defer pool.mu.RUnlock()
-	pods := []*ApplicationPod{}
-	for _, p := range pool.pods[state] {
-		pods = append(pods, p)
-	}
-	return pods
-}
-
-func (pool *ApplicationPool) podLengthOfState(podState ApplicationPodState) int {
-	pool.mu.RLock()
-	defer pool.mu.RUnlock()
-	return len(pool.pods[podState])
-}
-
-func (pool *ApplicationPool) podLength() int {
-	pool.mu.RLock()
-	defer pool.mu.RUnlock()
-	length := 0
-	for _, v := range pool.pods {
-		length += len(v)
-	}
-	return length
-}
-
-func (pool *ApplicationPool) podList() []*ApplicationPod {
-	pool.mu.RLock()
-	defer pool.mu.RUnlock()
-	pods := []*ApplicationPod{}
-	for _, v := range pool.pods {
-		for _, p := range v {
-			pods = append(pods, p)
-		}
-	}
-	return pods
-}
-
 func (am *ApplicationManager) onPodEventFromNode(podEvent *ie.PodEvent) {
 	klog.InfoS("Received a pod event", "pod", util.Name(podEvent.Pod), "type", podEvent.Type, "phase", podEvent.Pod.Status.Phase, "condition", k8spodutil.IsPodReady(podEvent.Pod))
 	if _, found := podEvent.Pod.Labels[fornaxv1.LabelFornaxCoreNodeDaemon]; !found {
