+++ pkg/util/meta.go
@@ -30,6 +30,13 @@ func NewCurrentMetaTime() *metav1.Time {
 	}
 }
 
+// NewCurrentMetaTimeNormallized return a metav1.Time initilized with time.Now, truncate to second
+func NewCurrentMetaTimeNormallized() *metav1.Time {
+	return &metav1.Time{
+		Time: time.Now().Truncate(time.Second),
+	}
+}
+
 // AddFinalizer modify passed object meta and add specified finalizer into its list,
 // if there is already specified finalizer, no change
 func AddFinalizer(ometa *metav1.ObjectMeta, finalizer string) {
